requirements.txt
pefile
python-magic
yara-python
virustotal-api
oletools
M2Crypto